LayoutTests/fast/mediastream/MediaDevices-getSupportedConstraints.html

<!DOCTYPE HTML>
<html>
    <head>
        <script src="../../resources/js-test-pre.js"></script>
    </head>
    <body>
        <p id="description"></p>
        <div id="console"></div>
        <script>
            var supportedConstraints;

            description("Tests MediaDevices.getSupportedConstraints()");
            window.jsTestIsAsync = true;

            shouldBeType("navigator.mediaDevices", "Object");
            shouldBeType("navigator.mediaDevices.getSupportedConstraints", "Function");

            debug("");
            evalAndLog("supportedConstraints = navigator.mediaDevices.getSupportedConstraints()");
            debug("");
            for (var constraint in supportedConstraints)
                shouldBeTrue(`supportedConstraints.${constraint}`);
            debug("");


            finishJSTest();

            window.successfullyParsed = true;
        </script>
        <script src="../../resources/js-test-post.js"></script>
    </body>
</html>
